Move validation functions to separate file
